+++ src/num/geom/metric/distance.rs
@@ -16,8 +16,24 @@ use crate::{Orientation, Position};
 ///   and represents how far apart two positions are.
 /// - Unlike [`Orientation`], `Distance` has **magnitude**
 ///   but no defined **orientation**.
+#[must_use]
+#[repr(transparent)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct Distance<T, const D: usize> {
     /// The component-wise separation in `D`-dimensional space.
     pub dim: [T; D],
+}
+crate::items! {
+    impl<T, const D: usize> From<[T; D]> for Distance<T, D> {
+        fn from(dim: [T; D]) -> Self { Self { dim } }
+    }
+    impl<T> From<(T, T)> for Distance<T, 2> {
+        fn from(dim: (T, T)) -> Self { Self { dim: [dim.0, dim.1] } }
+    }
+    impl<T> From<(T, T, T)> for Distance<T, 3> {
+        fn from(dim: (T, T, T)) -> Self { Self { dim: [dim.0, dim.1, dim.2] } }
+    }
+    impl<T> From<(T, T, T, T)> for Distance<T, 4> {
+        fn from(dim: (T, T, T, T)) -> Self { Self { dim: [dim.0, dim.1, dim.2, dim.3] } }
+    }
 }