+++ src/num/geom/mod.rs
@@ -4,12 +4,11 @@
 #![doc = crate::doc_!(modules: crate::num; geom: linear, metric, shape)]
 //
 
+pub mod metric; // Position, Distance, Extent, Stride...
+
 #[cfg(feature = "linear")]
 #[cfg_attr(nightly_doc, doc(cfg(feature = "linear")))]
 pub mod linear; // Vector*, Matrix*
-#[cfg(feature = "metric")]
-#[cfg_attr(nightly_doc, doc(cfg(feature = "metric")))]
-pub mod metric; // Position, Distance, Extent, Stride...
 #[cfg(feature = "shape")]
 #[cfg_attr(nightly_doc, doc(cfg(feature = "shape")))]
 pub mod shape; // Point, …
@@ -22,10 +21,10 @@ crate::items! { // structural access: _mods, _all
 
     // mod _mods {}
     mod _pub_mods { #![allow(unused)]
+        pub use super::metric::_all::*;
+
         #[cfg(feature = "linear")]
         pub use super::linear::_all::*;
-        #[cfg(feature = "metric")]
-        pub use super::metric::_all::*;
         #[cfg(feature = "shape")]
         pub use super::shape::_all::*;
     }